+++ module-system/sov-test-utils/src/runtime/mod.rs
@@ -35,7 +35,7 @@ use traits::{MinimalGenesis, PostTxHookRegistry};
 pub use wrapper::{TestRuntimeWrapper, WorkingSetClosure};
 
 // Constants used in the genesis configuration of the test runtime
-const MIN_USER_BOND: u64 = 100_000;
+const MIN_USER_BOND: u64 = 100_000_000;
 const MAX_ATTESTED_HEIGHT: u64 = 0;
 const LIGHT_CLIENT_FINALIZED_HEIGHT: u64 = 0;
 const ROLLUP_FINALITY_PERIOD: u64 = 1;
@@ -269,7 +269,7 @@ impl<M: Module, S: Spec> MessageType<M, S> {
                 msg,
                 config_value!("CHAIN_ID"),
                 PriorityFeeBips::ZERO,
-                10_000_000,
+                100_000_000,
                 nonce,
                 None,
             ),
@@ -590,7 +590,7 @@ mod test_rt {
             &[],
             sequencer_rollup_addr,
             SEQUENCER_ADDR.into(),
-            100_000,
+            100_000_000,
             "SovereignToken".to_string(),
             10_000_000_000,
         );